Move password into own package to prevent package cycle
