+++ auth/password/password.go
@@ -1,4 +1,4 @@
-package auth
+package password
 
 import "golang.org/x/crypto/bcrypt"
